@@ -19,36 +19,42 @@ classifiers = [
     "Programming Language :: Python :: 3 :: Only",
     "Topic :: Scientific/Engineering :: GIS",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    "numpy >= 1.22",
+    "numpy >= 1.24",
     "pyogrio >= 0.7.2",
     "packaging",
-    "pandas >= 1.4.0",
-    "pyproj >= 3.3.0",
+    "pandas >= 2.0.0",
+    "pyproj >= 3.5.0",
     "shapely >= 2.0.0",
 ]
 
 [project.optional-dependencies]
 all = [
     "psycopg-binary>=3.1.0",
-    "SQLAlchemy>=1.3",
+    "SQLAlchemy>=2.0",
     "geopy",
-    "matplotlib>=3.5.0",
-    "mapclassify",
+    "matplotlib>=3.7",
+    "mapclassify>=2.5",
     "xyzservices",
-    "folium",
+    "folium", # >=0.12 (implied minimum version, but older versions may work)
     "GeoAlchemy2",
-    "pyarrow>=8.0.0",
+    "pyarrow>=10.0.0",
     "scipy",
 ]
+
+# Minimum supported additional deps, not installed as part of `all` extra
+# (useful to track for bumping minimum dep versions)
+# fiona>=1.8.21
+# scipy>=1.10.0
+
 dev = [
     "pytest>=3.1.0",
     "pytest-cov",
     "pytest-xdist",
     "codecov",
-    "black",
     "pre-commit",
+    "ruff",
 ]
 
 
@@ -96,9 +102,6 @@ filterwarnings = [
     "ignore:Cannot set the CRS, falling back to None:UserWarning:geopandas",
 ]
 
-[tool.black]
-line-length = 88
-
 [tool.ruff]
 line-length = 88
 extend-exclude = ["doc/*", "benchmarks/*", "versioneer.py", "geopandas/_version.py"]
@@ -111,7 +114,7 @@ select = [
     "E",
     "W",
     # pyupgrade
-    # "UP",
+    "UP",
     # flake8-bugbear
     "B",
     # flake8-debugger
