@@ -67,7 +67,7 @@ def _df_to_geodf(df, geom_col="geom", crs=None, con=None):
     """
 
     if geom_col not in df:
-        raise ValueError("Query missing geometry column '{}'".format(geom_col))
+        raise ValueError(f"Query missing geometry column '{geom_col}'")
 
     if df.columns.to_list().count(geom_col) > 1:
         raise ValueError(
@@ -104,7 +104,7 @@ def _df_to_geodf(df, geom_col="geom", crs=None, con=None):
                         f"Trying epsg:{srid} as a fallback."
                     )
                     warnings.warn(warning_msg, UserWarning, stacklevel=3)
-                    crs = "epsg:{}".format(srid)
+                    crs = f"epsg:{srid}"
                 else:
                     if not spatial_ref_sys_df.empty:
                         auth_name = spatial_ref_sys_df["auth_name"].item()
@@ -116,7 +116,7 @@ def _df_to_geodf(df, geom_col="geom", crs=None, con=None):
                             f"Trying epsg:{srid} as a fallback."
                         )
                         warnings.warn(warning_msg, UserWarning, stacklevel=3)
-                        crs = "epsg:{}".format(srid)
+                        crs = f"epsg:{srid}"
 
     return GeoDataFrame(df, crs=crs, geometry=geom_col)
 
@@ -327,11 +327,11 @@ def _psql_insert_copy(tbl, conn, keys, data_iter):
     writer.writerows(data_iter)
     s_buf.seek(0)
 
-    columns = ", ".join('"{}"'.format(k) for k in keys)
+    columns = ", ".join(f'"{k}"' for k in keys)
 
     dbapi_conn = conn.connection
-    sql = 'COPY "{}"."{}" ({}) FROM STDIN WITH CSV'.format(
-        tbl.table.schema, tbl.table.name, columns
+    sql = (
+        f'COPY "{tbl.table.schema}"."{tbl.table.name}" ({columns}) FROM STDIN WITH CSV'
     )
     with dbapi_conn.cursor() as cur:
         # Use psycopg method if it's available
@@ -435,19 +435,13 @@ def _write_postgis(
             # Only check SRID if table exists
             if connection.dialect.has_table(connection, name, schema):
                 target_srid = connection.execute(
-                    text(
-                        "SELECT Find_SRID('{schema}', '{table}', '{geom_col}');".format(
-                            schema=schema_name, table=name, geom_col=geom_name
-                        )
-                    )
+                    text(f"SELECT Find_SRID('{schema_name}', '{name}', '{geom_name}');")
                 ).fetchone()[0]
 
                 if target_srid != srid:
                     msg = (
-                        "The CRS of the target table (EPSG:{epsg_t}) differs from the "
-                        "CRS of current GeoDataFrame (EPSG:{epsg_src}).".format(
-                            epsg_t=target_srid, epsg_src=srid
-                        )
+                        f"The CRS of the target table (EPSG:{target_srid}) differs "
+                        f"from the CRS of current GeoDataFrame (EPSG:{srid})."
                     )
                     raise ValueError(msg)
 
