@@ -273,10 +273,8 @@ class TestIO:
         out_geom = "the_geom"
         create_postgis(con, df_nybb, geom_col=orig_geom)
 
-        sql = """SELECT borocode, boroname, shape_leng, shape_area,
-                    {} as {} FROM nybb;""".format(
-            orig_geom, out_geom
-        )
+        sql = f"""SELECT borocode, boroname, shape_leng, shape_area,
+                    {orig_geom} as {out_geom} FROM nybb;"""
         df = read_postgis(sql, con, geom_col=out_geom)
 
         validate_boro_df(df)
@@ -337,7 +335,7 @@ class TestIO:
         create_spatialite(con, df_nybb)
         sql = (
             "SELECT ogc_fid, borocode, boroname, shape_leng, shape_area, "
-            'AsEWKB("{0}") AS "{0}" FROM nybb'.format(geom_col)
+            f'AsEWKB("{geom_col}") AS "{geom_col}" FROM nybb'
         )
         df = read_postgis(sql, con, geom_col=geom_col)
         validate_boro_df(df)
@@ -349,7 +347,7 @@ class TestIO:
         create_spatialite(con, df_nybb)
         sql = (
             "SELECT ogc_fid, borocode, boroname, shape_leng, shape_area, "
-            'ST_AsBinary("{0}") AS "{0}" FROM nybb'.format(geom_col)
+            f'ST_AsBinary("{geom_col}") AS "{geom_col}" FROM nybb'
         )
         df = read_postgis(sql, con, geom_col=geom_col)
         validate_boro_df(df)
@@ -381,7 +379,7 @@ class TestIO:
         # Write to db
         write_postgis(df_nybb, con=engine, name=table, if_exists="fail")
         # Validate
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
 
@@ -397,7 +395,7 @@ class TestIO:
         # Write to db
         write_postgis(df_nybb, con=engine, name=table, if_exists="fail")
         # Validate
-        sql = text('SELECT * FROM "{table}";'.format(table=table))
+        sql = text(f'SELECT * FROM "{table}";')
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
 
@@ -413,7 +411,7 @@ class TestIO:
             # Write to db
             write_postgis(df_nybb, con=con, name=table, if_exists="fail")
             # Validate
-            sql = text("SELECT * FROM {table};".format(table=table))
+            sql = text(f"SELECT * FROM {table};")
             df = read_postgis(sql, con, geom_col="geometry")
             validate_boro_df(df)
 
@@ -451,7 +449,7 @@ class TestIO:
         # Overwrite
         write_postgis(df_nybb, con=engine, name=table, if_exists="replace")
         # Validate
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
 
@@ -469,21 +467,17 @@ class TestIO:
         write_postgis(df_nybb, con=engine, name=table, if_exists="replace")
         write_postgis(df_nybb, con=engine, name=table, if_exists="append")
         # Validate
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         new_rows, new_cols = df.shape
 
         # There should be twice as many rows in the new table
         assert new_rows == orig_rows * 2, (
-            "There should be {target} rows,found: {current}".format(
-                target=orig_rows * 2, current=new_rows
-            ),
+            f"There should be {orig_rows * 2} rows,found: {new_rows}",
         )
         # Number of columns should stay the same
         assert new_cols == orig_cols, (
-            "There should be {target} columns,found: {current}".format(
-                target=orig_cols, current=new_cols
-            ),
+            f"There should be {orig_cols} columns,found: {new_cols}",
         )
 
     @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
@@ -507,7 +501,7 @@ class TestIO:
         )
         with engine.connect() as conn:
             target_srid = conn.execute(sql).fetchone()[0]
-        assert target_srid == 0, "SRID should be 0, found %s" % target_srid
+        assert target_srid == 0, f"SRID should be 0, found {target_srid}"
 
     @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
     def test_write_postgis_with_esri_authority(self, engine_postgis, df_nybb):
@@ -530,7 +524,7 @@ class TestIO:
         )
         with engine.connect() as conn:
             target_srid = conn.execute(sql).fetchone()[0]
-        assert target_srid == 102003, "SRID should be 102003, found %s" % target_srid
+        assert target_srid == 102003, f"SRID should be 102003, found {target_srid}"
 
     @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
     def test_write_postgis_geometry_collection(
@@ -546,14 +540,10 @@ class TestIO:
         write_postgis(df_geom_collection, con=engine, name=table, if_exists="replace")
 
         # Validate geometry type
-        sql = text(
-            "SELECT DISTINCT(GeometryType(geometry)) FROM {table} ORDER BY 1;".format(
-                table=table
-            )
-        )
+        sql = text(f"SELECT DISTINCT(GeometryType(geometry)) FROM {table} ORDER BY 1;")
         with engine.connect() as conn:
             geom_type = conn.execute(sql).fetchone()[0]
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
 
         assert geom_type.upper() == "GEOMETRYCOLLECTION"
@@ -575,11 +565,7 @@ class TestIO:
         )
 
         # Validate geometry type
-        sql = text(
-            "SELECT DISTINCT GeometryType(geometry) FROM {table} ORDER BY 1;".format(
-                table=table
-            )
-        )
+        sql = text(f"SELECT DISTINCT GeometryType(geometry) FROM {table} ORDER BY 1;")
         with engine.connect() as conn:
             res = conn.execute(sql).fetchall()
         assert res[0][0].upper() == "LINESTRING"
@@ -598,11 +584,7 @@ class TestIO:
         write_postgis(df_linear_ring, con=engine, name=table, if_exists="replace")
 
         # Validate geometry type
-        sql = text(
-            "SELECT DISTINCT(GeometryType(geometry)) FROM {table} ORDER BY 1;".format(
-                table=table
-            )
-        )
+        sql = text(f"SELECT DISTINCT(GeometryType(geometry)) FROM {table} ORDER BY 1;")
         with engine.connect() as conn:
             geom_type = conn.execute(sql).fetchone()[0]
 
@@ -625,17 +607,13 @@ class TestIO:
             chunksize=1,
         )
         # Validate row count
-        sql = text("SELECT COUNT(geometry) FROM {table};".format(table=table))
+        sql = text(f"SELECT COUNT(geometry) FROM {table};")
         with engine.connect() as conn:
             row_cnt = conn.execute(sql).fetchone()[0]
         assert row_cnt == 3
 
         # Validate geometry type
-        sql = text(
-            "SELECT DISTINCT GeometryType(geometry) FROM {table} ORDER BY 1;".format(
-                table=table
-            )
-        )
+        sql = text(f"SELECT DISTINCT GeometryType(geometry) FROM {table} ORDER BY 1;")
         with engine.connect() as conn:
             res = conn.execute(sql).fetchall()
         assert res[0][0].upper() == "LINESTRING"
@@ -651,7 +629,7 @@ class TestIO:
 
         table = "nybb"
         schema_to_use = "test"
-        sql = text("CREATE SCHEMA IF NOT EXISTS {schema};".format(schema=schema_to_use))
+        sql = text(f"CREATE SCHEMA IF NOT EXISTS {schema_to_use};")
         with engine.begin() as conn:
             conn.execute(sql)
 
@@ -659,9 +637,7 @@ class TestIO:
             df_nybb, con=engine, name=table, if_exists="replace", schema=schema_to_use
         )
         # Validate
-        sql = text(
-            "SELECT * FROM {schema}.{table};".format(schema=schema_to_use, table=table)
-        )
+        sql = text(f"SELECT * FROM {schema_to_use}.{table};")
 
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
@@ -677,7 +653,7 @@ class TestIO:
 
         table = "nybb"
         schema_to_use = "test"
-        sql = text("CREATE SCHEMA IF NOT EXISTS {schema};".format(schema=schema_to_use))
+        sql = text(f"CREATE SCHEMA IF NOT EXISTS {schema_to_use};")
         with engine.begin() as conn:
             conn.execute(sql)
 
@@ -686,11 +662,7 @@ class TestIO:
                 df_nybb, con=engine, name=table, if_exists="fail", schema=schema_to_use
             )
             # Validate
-            sql = text(
-                "SELECT * FROM {schema}.{table};".format(
-                    schema=schema_to_use, table=table
-                )
-            )
+            sql = text(f"SELECT * FROM {schema_to_use}.{table};")
 
             df = read_postgis(sql, engine, geom_col="geometry")
             validate_boro_df(df)
@@ -704,9 +676,7 @@ class TestIO:
             df_nybb, con=engine, name=table, if_exists="replace", schema=schema_to_use
         )
         # Validate
-        sql = text(
-            "SELECT * FROM {schema}.{table};".format(schema=schema_to_use, table=table)
-        )
+        sql = text(f"SELECT * FROM {schema_to_use}.{table};")
 
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
@@ -723,7 +693,7 @@ class TestIO:
         write_postgis(df_3D_geoms, con=engine, name=table, if_exists="replace")
 
         # Check that all geometries have 3 dimensions
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         assert list(df.geometry.has_z) == [True, True, True]
 
@@ -740,7 +710,7 @@ class TestIO:
         write_postgis(df_nybb, con=engine, name=table, if_exists="replace")
 
         # Check that the row order matches
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         assert df["BoroCode"].tolist() == correct_order
 
@@ -758,7 +728,7 @@ class TestIO:
         write_postgis(df_nybb, con=engine, name=table, if_exists="append")
 
         # Check that the row order matches
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
 
