@@ -1,6 +1,5 @@
 import itertools
 import warnings
-from packaging.version import Version
 
 import numpy as np
 import pandas as pd
@@ -221,13 +220,6 @@ class TestPointPlotting:
                 np.linspace(0, 0.0, 1.0, self.N), ax.collections[0].get_alpha()
             )
 
-    # TODO temporary skip for mpl 3.8.0.dev
-    # (https://github.com/matplotlib/matplotlib/issues/25162)
-    @pytest.mark.skipif(
-        Version(matplotlib.__version__) >= Version("3.8.0.dev")
-        and Version(matplotlib.__version__) < Version("3.8.0"),
-        reason="failing with matplotlib dev",
-    )
     def test_legend(self):
         with warnings.catch_warnings(record=True) as _:  # don't print warning
             # legend ignored if color is given.
@@ -1924,9 +1916,9 @@ def _check_colors(N, actual_colors, expected_colors, alpha=None):
     ), "Different lengths of actual and expected colors!"
 
     for actual, expected in zip(all_actual_colors, expected_colors):
-        assert actual == conv.to_rgba(expected, alpha=alpha), "{} != {}".format(
-            actual, conv.to_rgba(expected, alpha=alpha)
-        )
+        assert actual == conv.to_rgba(
+            expected, alpha=alpha
+        ), f"{actual} != {conv.to_rgba(expected, alpha=alpha)}"
 
 
 def _style_to_linestring_onoffseq(linestyle, linewidth):
@@ -1955,7 +1947,7 @@ def _get_ax(fig, label):
     for ax in fig.axes:
         if ax.get_label() == label:
             return ax
-    raise ValueError("no ax found with label {0}".format(label))
+    raise ValueError(f"no ax found with label {label}")
 
 
 def _get_colorbar_ax(fig):
