@@ -27,7 +27,7 @@ _NATURALEARTH_LOWRES = os.path.join(
 try:
     from unittest import mock
 except ImportError:
-    import mock  # noqa: F401
+    from unittest import mock  # noqa: F401
 
 
 def validate_boro_df(df, case_sensitive=False):
@@ -111,8 +111,8 @@ def create_postgis(con, df, srid=None, geom_col="geom"):
     # > createdb test_geopandas
     # > psql -c "CREATE EXTENSION postgis" -d test_geopandas
     if srid is not None:
-        geom_schema = "geometry(MULTIPOLYGON, {})".format(srid)
-        geom_insert = "ST_SetSRID(ST_GeometryFromText(%s), {})".format(srid)
+        geom_schema = f"geometry(MULTIPOLYGON, {srid})"
+        geom_insert = f"ST_SetSRID(ST_GeometryFromText(%s), {srid})"
     else:
         geom_schema = "geometry"
         geom_insert = "ST_GeometryFromText(%s)"
@@ -120,22 +120,18 @@ def create_postgis(con, df, srid=None, geom_col="geom"):
         cursor = con.cursor()
         cursor.execute("DROP TABLE IF EXISTS nybb;")
 
-        sql = """CREATE TABLE nybb (
+        sql = f"""CREATE TABLE nybb (
             {geom_col}   {geom_schema},
             borocode     integer,
             boroname     varchar(40),
             shape_leng   float,
             shape_area   float
-            );""".format(
-            geom_col=geom_col, geom_schema=geom_schema
-        )
+            );"""
         cursor.execute(sql)
 
         for i, row in df.iterrows():
-            sql = """INSERT INTO nybb VALUES ({}, %s, %s, %s, %s
-            );""".format(
-                geom_insert
-            )
+            sql = f"""INSERT INTO nybb VALUES ({geom_insert}, %s, %s, %s, %s
+            );"""
             cursor.execute(
                 sql,
                 (
