@@ -245,9 +245,7 @@ def overlay(df1, df2, how="intersection", keep_geom_type=None, make_valid=True):
     ]
     # Error Messages
     if how not in allowed_hows:
-        raise ValueError(
-            "`how` was '{0}' but is expected to be in {1}".format(how, allowed_hows)
-        )
+        raise ValueError(f"`how` was '{how}' but is expected to be in {allowed_hows}")
 
     if isinstance(df1, GeoSeries) or isinstance(df2, GeoSeries):
         raise NotImplementedError(
@@ -271,9 +269,7 @@ def overlay(df1, df2, how="intersection", keep_geom_type=None, make_valid=True):
         lines_check = df.geom_type.isin(lines).any()
         points_check = df.geom_type.isin(points).any()
         if sum([poly_check, lines_check, points_check]) > 1:
-            raise NotImplementedError(
-                "df{} contains mixed geometry types.".format(i + 1)
-            )
+            raise NotImplementedError(f"df{i + 1} contains mixed geometry types.")
 
     if how == "intersection":
         box_gdf1 = df1.total_bounds
@@ -353,9 +349,7 @@ def overlay(df1, df2, how="intersection", keep_geom_type=None, make_valid=True):
             elif geom_type in points:
                 exploded.loc[~exploded.geom_type.isin(points), geom_col] = None
             else:
-                raise TypeError(
-                    "`keep_geom_type` does not support {}.".format(geom_type)
-                )
+                raise TypeError(f"`keep_geom_type` does not support {geom_type}.")
             num_dropped_collection = (
                 orig_num_geoms_exploded - exploded.geometry.isna().sum()
             )
@@ -382,15 +376,15 @@ def overlay(df1, df2, how="intersection", keep_geom_type=None, make_valid=True):
         elif geom_type in points:
             result = result.loc[result.geom_type.isin(points)]
         else:
-            raise TypeError("`keep_geom_type` does not support {}.".format(geom_type))
+            raise TypeError(f"`keep_geom_type` does not support {geom_type}.")
         num_dropped = orig_num_geoms - result.shape[0]
 
         if (num_dropped > 0 or num_dropped_collection > 0) and keep_geom_type_warning:
             warnings.warn(
-                "`keep_geom_type=True` in overlay resulted in {} dropped "
-                "geometries of different geometry types than df1 has. "
-                "Set `keep_geom_type=False` to retain all "
-                "geometries".format(num_dropped + num_dropped_collection),
+                "`keep_geom_type=True` in overlay resulted in "
+                f"{num_dropped + num_dropped_collection} dropped geometries of "
+                "different geometry types than df1 has. Set `keep_geom_type=False` to "
+                "retain all geometries",
                 UserWarning,
                 stacklevel=2,
             )
