@@ -21,18 +21,16 @@ def test_no_additional_imports():
         "matplotlib",
     }
 
-    code = """
+    code = f"""
 import sys
 import geopandas
-blacklist = {0!r}
+blacklist = {blacklist!r}
 
 mods = blacklist & set(m.split('.')[0] for m in sys.modules)
 if mods:
     sys.stderr.write('err: geopandas should not import: {{}}'.format(', '.join(mods)))
     sys.exit(len(mods))
-""".format(
-        blacklist
-    )
+"""
     call = [sys.executable, "-c", code]
     returncode = subprocess.run(call, check=False).returncode
     assert returncode == 0
