@@ -1011,31 +1011,27 @@ class TestGeomMethods:
             [
                 Polygon(
                     (
-                        (
-                            (0, 0),
-                            (0.5, 0),
-                            (1, 0),
-                            (1, 0.5),
-                            (1, 1),
-                            (0.6666666666666666, 0.6666666666666666),
-                            (0.3333333333333333, 0.3333333333333333),
-                            (0, 0),
-                        )
+                        (0, 0),
+                        (0.5, 0),
+                        (1, 0),
+                        (1, 0.5),
+                        (1, 1),
+                        (0.6666666666666666, 0.6666666666666666),
+                        (0.3333333333333333, 0.3333333333333333),
+                        (0, 0),
                     )
                 ),
                 Polygon(
                     (
-                        (
-                            (0, 0),
-                            (0.5, 0),
-                            (1, 0),
-                            (1, 0.5),
-                            (1, 1),
-                            (0.5, 1),
-                            (0, 1),
-                            (0, 0.5),
-                            (0, 0),
-                        )
+                        (0, 0),
+                        (0.5, 0),
+                        (1, 0),
+                        (1, 0.5),
+                        (1, 1),
+                        (0.5, 1),
+                        (0, 1),
+                        (0, 0.5),
+                        (0, 0),
                     )
                 ),
             ]
